Add clang-format target to CMake, integrate clang-tidy in CMake
format-cxx is formats the codebase according to the .clang-tidy file

clang-tidy is now used in CMake, with the compile options ELSA_CLANG_TIDY, it can be turned on. Then building elsa will show the clang-tidy checks as well. CI pipeline uses this target with a new build-tests target, which only builds the tests and does not run it. The job exports a clang-tidy-fix.txt file, which shows possible fixes for the stuff clang-tidy found