Reworked CI caching; Reworked CMake submodule update routine
- Chaching is introduced for for all jobs, it caches for a branch.
only submodules are chached, as build folders should be cached using
artifacts, which is also done. We also cache the .git/modules folder
as anyway we have to reclone everything everytime.
- CMake UpdateGitSubmodules now goes over all submodules and check
individually if they have to be initialized, or just updated.
Also work around a specific error in the CI pipeline, if
.git/modules/thirdparty/<submodule> is not present (which by default
dosent get cached), we have to deinit it and then reinit it.