Express intent of members of LinearResidual better
Instead of using extra bool flags to indicate whether or not
`LinearResidual` has an operator or data vector, use the type system to
better express intent.

A non existing operator is represented with an null pointer. And the
data vector is wrapped in an optional, which first is clearer in intent
but also removed the need for an unnecessary heap allocation.