Use CMAKE_CUDA_ARCHITECTURES to determine compute capabilities
In CMake 3.18 the (CMAKE_)CUDA_ARCHITECTURES variables were introduced.
CMake uses them to pass the compute capabilities to the CUDA compiler.

Therefore, for CMake versions 3.18 and higher, we rely on this CMake
behaviour and while we still support previous versions, the previous
behaviour is still kept around.

CMAKE_CUDA_ARCHITECTURES is set to 'Auto' by default, which is
equivalent to the previous CUDA_ARCH variables that was used.
Overwriting via the command line is possible for cross compilation of
CUDA code.