Reduce scope of caching in CI to a cache per job and branch
As found in MR !153, the scope of the caches in the CI were chosen to
large. Therefore it could happen, that sporadically a build job, would
pick a cache from the clang-tidy run, which initializes with the
"Makefile" generator rather than Ninja. This would cause the pipeline to
fail.

Now, the caching is more conservative, there is one cache for each
job-branch combination. This should completely resolve this issue, but
it's also not optimal from a space requirement, but I'd rather use more
disk space on the build server, than have sporadic failures of the CI.