Remove unnecessray heap allocation in ml
At a couple of places a new unique_ptr was created, then directly
dereferenced and a copy of the underlying VolumeDescriptor returned.
This creates an unnecessray heap allocation. This commits directly casts
the reference of the smart pointer to a VolumeDescriptor and from there
the copy is created.

Remove dynamic_unique_ptr_cast and static_unique_ptr_cast as both is
covered by the downcast functions in TypeCasts.hpp.