#124 Use self-hosted docker registry to pull images in CI
This commit changes the pipelines to use the newly created docker
registry. The docker registry is running at vmlasser7.in.tum.de with
port 5000. You'll need the credentials and run 'docker login
docker.ciip.in.tum.de:5000' to pull and push images to the registry.

A new variable 'DOCKER_AUTH_CONFIG' is set up in GitLab to manage the
authentication inside of the CI pipeline.