Remove noise from Timer in logging output
The Timer class now prints timing information only if the Logger is set
to debug level, instead of info level. This reduces noise in the output.
The Timing of execution is not necessarily an useful information all the
time.

An example is the python script introduced in !182, all useful
information regarding conversion and solvers is shadowed by the Timing
information.