#101 Update more tests to be idiomatic
This is a follow up to !142, more tests use the assertion macros with
*_EQ, *_UNARY and such. This commit also fixes many warnings in the
tests with conversions.

Further warnings are fixed in many places such as the CUDA projectors,
the Dictionary operator or the SQS solver. As the last thing, this
commits adds a new overload of isApprox for Expressions.

With this commit, elsa basically without warnings on GCC 11 and Clang 12