#105 Add slicing capabilities to DataContainer
Access to a portion of the DataContainer in the form of slices.
Currently, only slicing in the direction of the last dimension are
possible, but this should still be useful for many cases.

The one suprising thing in this implementation, is that the returned
DataContainer, has a descriptor which is of the same dimension as the
original DataContainer but with a "thickness" of 1 in the last
dimension, instead of actually being lower dimensional.