Use ctests junit reports to improve error output in CI
With this commit, all test binaries use regular console output, and
ctests generates the junit xml report used for nice visual presentation
in GitLab (which was broken since I don't know when, so it should work
again now). Previously, no output was generated, as it was piped to the
test binaries unit report, which made finding errors unnecessary
complex.

Failed tests are rerun again such that the output of these tests is
found at the bottom of the CI log, which should make finding errors a
little easier in the future.