#146 Update tests and quickvec to fix pipeline
This commit tells CPM to fetch a specific commit of quickvec instead of
the master branch. Then new changes have to be added via a MR in elsa
and therefore the whole pipeline will be run. This should prevent errors
such as #146.

Also tests for CG are adapted, for whatever reason they failed, now they
don't anymore...