Reduce spdlog version 1.8.5
To build all docker images, for some reason it's necessary to reduce the
spdlog version from 1.9.0 to 1.8.5. For not known reasons, the build for
the GPU only image fails with spdlog version 1.9.0, but it somehow
succeeds for the current pipeline image. I don't know why that is and
how to fix it, so for now reduce it, that images can still be updated
with new CMake versions and so on.