docs: Fix build deps and Python Bindings page.
[CONTRIBUTING.md]
- In order to be able to build the documentation, the package
`sphinxcontrib-katex` also needs to be installed via pip.

[docs/guides/python_bindings.md]
- According to the matplotlib documentation [1], the second argument
to `imshow` needs to be a string describing a colormap name or an
instance of a ColorMap. An error occurred with the previously used
argument as it did not describe a defined colormap. As the argument is
optional, it can just be omitted (or replaced according to personal
style preferences).
- The relative link to the Quickstart page did not work, it has been
replaced by a doc-link according to the sphinx documentation [2].

[1] matplotlib.org/3.5.0/api/_as_gen/matplotlib.pyplot.imshow.html
[2] docs.readthedocs.io/en/stable/guides/cross-referencing-with-sphinx.html#the-doc-role