refactor: remove default iterations from derived solvers
This removes currently unused default number of iterations in each
derived solver. Now the base class defaults the number of iterations to
the value. This is the same value for all solvers anyway. If this is an
required feature for certain solvers, this can be added on a case by
case basis.