chore: remove expression templates
Handling heterogeneous data on CPU and GPU is painful. We decided to
explore a different direction using unified memory, if CUDA support is
enabled, else just rely on a std::vector like abstraction. However, we
are limited by our DataHandler design, which needs to change completely
for this to work out properly. Hence, expression templates is the first
part which will need to go...

It will be a future project to add them back in a way, which isn't as
intrusive to the storage as it is now.