refactor: factor out composition logic from LinearOperator
The LinearOperator so far did very many tasks, among others:
- Composition of `scalar * operator`
- Composition of `operator * operator`
- Composition of `operator + operator`
- Dealing with adjoint operators

To make it a little easier to digest, each case is now a new class
instead of cluttering it all in the LinearOperator. This hopefully eases
the uses a little. However, this made the LinearOperator a pure virtual
class and hence `getHessian` and `getJacobian` now return a unique_ptr
to an operator, which is meeh