feat: speed up ellipsoid rasterization
To speed up the rasterization of the shepp logan phantom
two new methods handle the rasterization of rotated and not
rotated ellipsoids. They are called via elsa::phantoms::rasterize().
To do so, a new class Ellipsoid is introduced.

To compare the old and the new versions for my BT the namespace
elsa::phantoms::old contains the old rasterization algorithm.