refactor: Replace inheritance in ProximityOperator to type erasure
Type erasure is a method to achieve runtime polymorphism, similar to
classical C++ inheritance. However, it has a couple of nice aspects.

The main difference is, that any class can "bind" to the base class, as
long as it adheres to the interface. This mostly means, that as long as
a class doesn't bind to the base class, it's an independend class.
Without the cost of virtual functions, complicated cloning and such and
such. Hence, it improves local reasoning of types. You only need to know
about the interface, when you want to "bind" against it, e.g. as a
function parameter.

The one major drawback of type erasure is the rather complicated
implementation of the base class. However, this only needs to be done
for interfaces, and so it's rather rare.

This certainly is an experiment, however, I'm quite confident, that it
will improve our codebase overall (due to an increased local reasoning,
most types are just regular and such).

Proximal operators are currently still rather small, and hence it is a
good playground without ripping apart the whole framework. We can try it
here, and see if we want to further use it.