refactor: remove Residual from functionals
In the current design, the `Functional` base class contains a pointer to
some `Residual`. This tends to be an over generalization, especially
with the further focus on proximal methods. For many functionals, the
proximal operator is only known for specific and simple versions of the
residual. Further, this generalizeation didn't really provide use with
much, the only interesting residual is the linear residual.

In the future, we will introduce more simple functionals, which do
exactly one thing and have an implicit residual. I.e. split the current
`L2NormPow2` into the least squares functional, with a full linear
residual, and the `L2NormSquared` which can only contain a translation.