refactor: remove problems in favour of general functionals
We decided to remove the current design based on explicit `Problem`
classes, mostly due to future (anticipated) sclaing issues.

There are mostly two categories of reconstruction algorithms in elsa
(very much simplified of course). The ones that can only handle a
specific form of problem (i.e. only Least Squares, (generalized)
Tikhonov, LASSO and so on), or expect some certain properties of the
problem (i.e. once differentiable).

For the second class, we so far have no representation at all. But for
those, the `Problem` base class, only acted as a sum of functionals.

For the first class, the abstraction made sense. But for more advanced
reconstruction frameworks such as proximal gradient descent, ADMM or
PDHG this is problematic. There might not be a unique mapping from a
problem to the framework. A simple example is a least squares problem
with a mix of L1 and L2 regularization using proximal gradient descent
(ISTA or FISTA). The L2 term is differentiable, hence it could both be
part of the differentiable function `f`, but a simple proximal mapping
is also known, hence you can use it in `g` in combination with the L1
regularization.

So conversion from one problem to another become highly non-trivial. We
decided to drop it, in favour of each solver expecting whatever it needs
to solve their problem. Some more work is necessary to make this as easy
and convenient as possible. But in general, less generalized knowledge
should (hopefully) be necessary to use specific reconstruction
algorithms.