refactor: rework ADMM to be easier to use
A couple of things have been a little a pain point. The ADMM version
used templates, instead of our solver interface, which made it sometimes
a little clunky to use. Further, it was limited to L1 and L0
regularization.

Now it is generalized to a version, where `f` is the least squares
functional and is solved explicitly, and `g` need to be some proximal.
Some tricks that where used previously, are removed and will return in a
future version.