feat: add simple implementation of CG
I've run into a couple of problems with CGLS with the
JosephsProjectorCUDA. As that is unmatched, the algorithm doesn not
converge properly. This took me quite some time to figure out, and in
the meantime I've implemented CG again to check if I'm stupid, so now we
have it as again. And it seems to be more stable anyway.