feat: add function for linear combination
Many iterative algorithms - e.g variants of CG or ADMM - require
expressions of the form `a * x + b * y`, where `a` and `b` are scalars,
and `x` and `y` vectors.

Evaluating the expression in the naive manner requires many allocations
and hence can be slower compared to the now implemented memory efficient
version.

After some quick microbenchmarks, under the assumption of preallocated
data, i.e. `lincomb(a, x, b, y, out)` vs `out = a * x + b * y`, the
speedup is around 2-6x. It seems especially noticeable the larger the
volumes get. The benchmark code is found in the benchmark folder.