feat: add point-wise L1 and L2 norm to DataContainer
The point-wise norms assume a blocked DataContainer, and compute the
respective norm column-wise. Meaning, that you form a matrix, where
each block represents a row of the matrix, and the norm is computed
along the column. The return type of the point-wise norm is again
a DataContainer.

This norm is useful for isotropic and anisotropic TV regularizations.