temporary performance fix
The observed performance degradation was due to unnecessary copies being performed (on the host) during move assignment.
Copies are performed during moves, whenever the containers are using different memory resources.
The moved to container previously did not take over the resource used by the moved from container.
With this commit, it temporarily does. The underlying problem to investigate is: why are different
memory resources returned by baseline instance? The baseline instance should never be set in our
example. The other thing fixed by this commit: no longer perform host copies in move constructors.
This was due to the use of a raw, untagged pointer type in ContiguousVector::assign(...).